Add temporary column course_duration with integer format

+++ database/migrations/2024_05_02_113700_add_price_column_into_courses_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('courses', function (Blueprint $table) {
-            $table->float('price', 2, 1)->after('hourly_rate');
+            $table->integer('price')->after('hourly_rate');
         });
     }
 };
